read_json method write_configuration_file method
