+++ src/ansys/aedt/core/rmxprt.py
@@ -380,7 +380,7 @@ class Rmxprt(FieldAnalysisRMxprt):
             ``True`` when successful, ``False`` when failed.
         """
 
-        from ansys.aedt.core.generic.general_methods import read_configuration_file
+        from ansys.aedt.core.generic.file_utils import read_configuration_file
 
         new_dict = read_configuration_file(input_file)
         for k, v in new_dict.items():
