+++ src/ansys/aedt/core/workflows/hfss/choke_designer.py
@@ -27,7 +27,7 @@ import tempfile
 
 import ansys.aedt.core
 from ansys.aedt.core import Hfss
-from ansys.aedt.core.generic.general_methods import read_json
+from ansys.aedt.core.generic.file_utils import read_json
 from ansys.aedt.core.generic.general_methods import write_configuration_file
 import ansys.aedt.core.workflows
 from ansys.aedt.core.workflows.misc import get_aedt_version
