+++ tests/system/general/test_12_1_PostProcessing.py
@@ -27,8 +27,8 @@ import sys
 import uuid
 
 from ansys.aedt.core import Quantity
+from ansys.aedt.core.generic.file_utils import read_json
 from ansys.aedt.core.generic.general_methods import is_linux
-from ansys.aedt.core.generic.general_methods import read_json
 from ansys.aedt.core.generic.settings import settings
 from ansys.aedt.core.visualization.plot.pyvista import _parse_aedtplt
 from ansys.aedt.core.visualization.plot.pyvista import _parse_streamline
