+++ tests/system/general/test_20_HFSS.py
@@ -1313,7 +1313,7 @@ class TestClass:
     )
     def test_51a_array(self):
         self.aedtapp.insert_design("Array_simple", "Modal")
-        from ansys.aedt.core.generic.general_methods import read_json
+        from ansys.aedt.core.generic.file_utils import read_json
 
         if config["desktopVersion"] > "2023.1":
             dict_in = read_json(
